move auth and profile files into the routes directory
